fix: refactor to shared utils
