+++ packages/angular-mcp-server/src/lib/tools/ds/story-parser/utils/story-parser.utils.ts
@@ -1,12 +1,21 @@
 import * as ts from 'typescript';
 import {
+  extractTemplateLiteral,
+  extractBalancedBlock,
+} from '@push-based/typescript-ast-utils';
+import {
   StoryFileData,
-  SelectorInfo,
   ArgTypeEntry,
   MetaArg,
   StoryEntry,
   ArgsOverride,
 } from '../models/types.js';
+import {
+  cleanTemplate,
+  extractSlotsFromTemplates,
+  detectSelectorStyle,
+  detectFormIntegration,
+} from '../../shared/utils/template-helpers.js';
 
 // ============================================================================
 // Main entry point
@@ -877,264 +886,4 @@ export function extractTernaryElseBranch(storyBody: string): string | null {
   if (backtickStart === -1) return null;
 
   return extractTemplateLiteral(storyBody, backtickStart);
-}
-
-/**
- * Extract content of a template literal starting at the opening backtick.
- * Handles nested ${} expressions.
- */
-function extractTemplateLiteral(
-  source: string,
-  openIdx: number,
-): string | null {
-  let i = openIdx + 1; // skip opening backtick
-  let depth = 0;
-  let result = '';
-
-  while (i < source.length) {
-    if (depth === 0 && source[i] === '`') {
-      return result;
-    }
-
-    if (source[i] === '$' && i + 1 < source.length && source[i + 1] === '{') {
-      depth++;
-      result += '${';
-      i += 2;
-      continue;
-    }
-
-    if (depth > 0 && source[i] === '{') {
-      depth++;
-    }
-
-    if (depth > 0 && source[i] === '}') {
-      depth--;
-      if (depth === 0) {
-        result += '}';
-        i++;
-        continue;
-      }
-    }
-
-    // Inside a nested expression, skip inner backtick strings
-    if (depth > 0 && source[i] === '`') {
-      i++;
-      let innerDepth = 0;
-      while (i < source.length) {
-        if (innerDepth === 0 && source[i] === '`') {
-          i++;
-          break;
-        }
-        if (
-          source[i] === '$' &&
-          i + 1 < source.length &&
-          source[i + 1] === '{'
-        ) {
-          innerDepth++;
-          i += 2;
-          continue;
-        }
-        if (innerDepth > 0 && source[i] === '}') {
-          innerDepth--;
-          i++;
-          continue;
-        }
-        if (source[i] === '\\') i++;
-        i++;
-      }
-      continue;
-    }
-
-    // Inside a nested expression, skip quoted strings
-    if (depth > 0 && (source[i] === "'" || source[i] === '"')) {
-      const q = source[i];
-      i++;
-      while (i < source.length && source[i] !== q) {
-        if (source[i] === '\\') i++;
-        i++;
-      }
-      i++; // skip closing quote
-      continue;
-    }
-
-    result += source[i];
-    i++;
-  }
-
-  return result || null;
-}
-
-/**
- * Extract balanced { } block starting at the opening brace position.
- */
-function extractBalancedBlock(source: string, openIdx: number): string | null {
-  let depth = 0;
-  let i = openIdx;
-
-  while (i < source.length && source[i] !== '{') i++;
-  if (i >= source.length) return null;
-
-  const start = i;
-  for (; i < source.length; i++) {
-    if (source[i] === '{') depth++;
-    else if (source[i] === '}') {
-      depth--;
-      if (depth === 0) return source.slice(start, i + 1);
-    }
-
-    // Skip string literals
-    if (source[i] === '`') {
-      i++;
-      while (i < source.length && source[i] !== '`') {
-        if (source[i] === '\\') i++;
-        i++;
-      }
-    } else if (source[i] === "'" || source[i] === '"') {
-      const quote = source[i];
-      i++;
-      while (i < source.length && source[i] !== quote) {
-        if (source[i] === '\\') i++;
-        i++;
-      }
-    }
-  }
-  return null;
-}
-
-/**
- * Clean a template string:
- * - Replace ${...} interpolations with `...` placeholders
- * - Remove Storybook layout wrapper divs
- * - Normalize whitespace
- * Returns null if the cleaned result doesn't look like valid HTML
- * (e.g., programmatically generated templates with JS fragments).
- */
-export function cleanTemplate(template: string): string | null {
-  let t = template;
-
-  // Replace argsToTemplate(args) with a meaningful marker before generic replacement
-  t = t.replace(/\$\{argsToTemplate\([^)]*\)\}/g, '[all-meta-args-bound]');
-
-  // Replace ${...} interpolations with placeholder
-  t = t.replace(/\$\{[^}]*\}/g, '...');
-
-  // Remove HTML comments
-  t = t.replace(/<!--[\s\S]*?-->/g, '');
-
-  // Remove ALL inline style attributes (storybook layout concerns, not component API)
-  t = t.replace(/\s*style=["'][^"']*["']/gi, '');
-
-  // Remove storybook layout wrapper divs (with class but layout-only) — strip opening AND closing tag
-  t = t.replace(
-    /<div\s+class="[^"]*(?:example-|tabgroup-)[^"]*">([\s\S]*?)<\/div>/gi,
-    '$1',
-  );
-
-  // Collapse to single line: strip all newlines and excess whitespace
-  t = t.replace(/\s+/g, ' ').trim();
-
-  // Remove whitespace between tags (but not inside tags)
-  t = t.replace(/>\s+</g, '><');
-
-  // Detect programmatically generated templates (JS fragments, not HTML)
-  if (!looksLikeHtml(t)) {
-    return null;
-  }
-
-  return t;
-}
-
-/**
- * Check if a cleaned template string looks like valid HTML rather than
- * JS code fragments from programmatically generated templates.
- */
-function looksLikeHtml(template: string): boolean {
-  // Must contain at least one HTML tag
-  if (!/<\w/.test(template)) return false;
-
-  // JS artifacts that indicate a programmatic template
-  const jsArtifacts = ['.join(', '.map(', '=>', 'function ', 'return '];
-  const artifactCount = jsArtifacts.filter((a) => template.includes(a)).length;
-  if (artifactCount >= 2) return false;
-
-  return true;
-}
-
-// ============================================================================
-// Slot extraction
-// ============================================================================
-
-/**
- * Scan templates for slot="name" patterns, deduplicate and sort alphabetically.
- */
-export function extractSlotsFromTemplates(templates: string[]): string[] {
-  const slots = new Set<string>();
-  const slotRegex = /slot="([^"]+)"/g;
-
-  for (const template of templates) {
-    let match: RegExpExecArray | null;
-    while ((match = slotRegex.exec(template)) !== null) {
-      slots.add(match[1]);
-    }
-  }
-
-  return [...slots].sort();
-}
-
-// ============================================================================
-// Selector style detection
-// ============================================================================
-
-/**
- * Detect whether a component uses attribute-style or element-style selectors.
- */
-export function detectSelectorStyle(
-  content: string,
-  kebabName: string,
-): SelectorInfo {
-  // Check for attribute usage: <button ds-{name}> or <a ds-{name}>
-  const attrRegex = new RegExp(
-    `<(?:button|a)\\s[^>]*\\bds-${escapeRegex(kebabName)}\\b`,
-    'i',
-  );
-  if (attrRegex.test(content)) {
-    return {
-      style: 'attribute',
-      selector: `ds-${kebabName}`,
-      note: 'Applied as attribute on `<button>` or `<a>`',
-    };
-  }
-
-  // Check for element usage: <ds-{name} followed by space, >, or newline
-  const elemRegex = new RegExp(
-    `<ds-${escapeRegex(kebabName)}(?=[\\s>/\\[])`,
-    'i',
-  );
-  if (elemRegex.test(content)) {
-    return { style: 'element', selector: `ds-${kebabName}` };
-  }
-
-  return { style: 'unknown', selector: `ds-${kebabName}` };
-}
-
-function escapeRegex(str: string): string {
-  return str.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-// ============================================================================
-// Form integration detection
-// ============================================================================
-
-/**
- * Detect Angular form integration patterns in templates.
- */
-export function detectFormIntegration(content: string): string[] {
-  const patterns: string[] = [];
-  if (/\[\(ngModel\)\]/.test(content))
-    patterns.push('ngModel (template-driven)');
-  if (/formControlName/.test(content))
-    patterns.push('formControlName (reactive forms)');
-  if (/formControl[^N]/.test(content))
-    patterns.push('formControl (reactive forms)');
-  return patterns;
 }
