+++ packages/shared/typescript-ast-utils/src/index.ts
@@ -1,5 +1,10 @@
 export * from './lib/constants.js';
 export * from './lib/utils.js';
+export * from './lib/source-text.utils.js';
 
 export { removeQuotes } from './lib/utils.js';
 export { QUOTE_REGEX } from './lib/constants.js';
+export {
+  extractTemplateLiteral,
+  extractBalancedBlock,
+} from './lib/source-text.utils.js';
