+++ packages/shared/utils/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/utils.js';
 export * from './lib/execute-process.js';
 export * from './lib/logging.js';
+export * from './lib/regex.utils.js';
 export * from './lib/file/find-in-file.js';
 export * from './lib/file/file.resolver.js';
 export * from './lib/file/default-export-loader.js';
